requirements.txt
tenacity>=8.1.0,<10.0.0
gradio>=5.5.0
omegaconf==2.2.3
mega.py
pycryptodome
tensorboardX
faiss-cpu==1.7.3
torchcrepe==0.0.23
ffmpeg-python>=0.2.0
einops
local-attention
edge-tts
pydub
matplotlib==3.7.1
asteroid==0.7.0
pyloudnorm
wandb
praat-parselmouth
webrtcvad
pandas
scipy
ml_collections
tqdm
torch_audiomentations
auraloss
torchseg
bitsandbytes
accelerate
huggingface-hub
prodigyopt
torch_log_wmse
audio-separator[gpu]==0.30.0